requirements.txt
langchain-core
langgraph
mcp[cli]
langchain-mcp-adapters
langchain-ollama
streamlit
typing-extensions
pydantic
paho-mqtt